refactor: unify ledger display flow
Ledger Nano SDK v.160 allows the NanoS and NanoX to share the same UX flow pattern.

This PR unifies the code which handles the UX display for NanoS and NanoX.

Specifically, this PR proposes the following:
- replace `src/ux` with a `src/display` directory.
- move and rename `src/ux/display_context.h/c` -> `src/display/context.h/c`.
- create a preprocessor header(`display/display.h`) to detect the nano platform.
- unify `src/ux/nanos` and `src/ux/nanox` to `src/display/nano.h/c`.
- remove legacy display code from `src/io.c`.
- update headers in all relevant files.
- update the changelog.

+++ src/operations/transactions/ux/display_ux.c
@@ -32,8 +32,8 @@
 #include "transactions/ux/htlc_claim_ux.h"
 #include "transactions/ux/htlc_refund_ux.h"
 
-#include "ux.h"
-#include "ux/display_context.h"
+#include "display/context.h"
+#include "display/display.h"
 
 ////////////////////////////////////////////////////////////////////////////////
 
@@ -48,7 +48,7 @@ void setDisplay(const Transaction *transaction) {
         case TRANSFER_TYPE:
             displayTransfer(transaction);
             setDisplaySteps(STEPS_TRANSFER_MIN +
-                            (uint8_t)(transaction->vendorFieldLength != 0U));
+                            transaction->vendorFieldLength != 0U);
             break;
 
         case SECOND_SIGNATURE_TYPE:
